🐛 Set debhelper-compat to version supported for all builds
Since we're currently building packages from Ubuntu 14.04 up to 23.04
it is the easiest and most stable way to depend on a debhelper-compat
version that is availabel on all releases. This is ATM version 9.

While this yields a linter warning it is IMO preferred to adjusting
the Build-Depends line automatically during the builds for the older
Ubuntu releases. Relying on an up-to-date version would perhaps
motivate at some point to use features only available (or, worse, with
changed outcome) with newer debhelper versions. Using the same
versiona across all builds should preserve us from such problems.